backend-data/requirements.txt
playwright>=1.40.0
requests>=2.28.0
flask>=3.0.0
flask-cors>=4.0.0
pandas>=2.0.0
openpyxl>=3.1.0
gspread>=5.12.0
python-dotenv>=1.0.0
supabase>=2.0.0
